make utility process available as API
